requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
yt-dlp==2024.11.18
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
opencv-python==4.8.1.78
numpy==1.24.3
openai==1.57.4
python-dotenv==1.2.1
